refactor: change entry point from index to www
