+++ bin/www.ts
@@ -1,0 +1,14 @@
+import express, { Express } from "express";
+import bodyParser from 'body-parser'
+import dotenv from "dotenv"
+
+const app: Express = express()
+dotenv.config()
+app.use(bodyParser.json({limit: "100mb"}))
+app.use(bodyParser.urlencoded({limit:"50mb", extended: true}))
+
+const port = process.env.PORT || 5000
+
+app.listen(port, () => {
+    console.log(`Server running on port ${port}`)
+})
