+++ src/http/bootstrap.ts
@@ -1,0 +1,6 @@
+import express from "express";
+import router from "./routes/routes";
+
+const app = express();
+
+app.use("/", router);
